fixed pointer bug

+++ printf/pull.c
@@ -6,7 +6,7 @@
 /*   By: bchan <marvin@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/02/07 12:31:26 by bchan             #+#    #+#             */
-/*   Updated: 2018/02/19 12:39:11 by bchan            ###   ########.fr       */
+/*   Updated: 2018/02/19 13:02:11 by bchan            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -85,7 +85,7 @@ char	*pull_voidp(t_print *form, va_list ap)
 	char	*temp;
 
 	arg = va_arg(ap, void *);
-	temp = ft_hex_convert((unsigned int)arg);
+	temp = ft_hex_convert((uintmax_t)arg);
 	result = ft_strjoin("0x", temp);
 	free(temp);
 	result = modify_string(form, result);
